Make the file enumeration callback based
To test how Hörbär's find_media enumeration, which is also callback based, could
be integrated into a flutter app.

To keep the iterative interface, a background thread runs the enumeration, and a
blocking fixed-capacity SPSC queue is used to transfer the entries to the
calling thread. (ideally this

For simplicity, this implementation spawns a new thread for each enumeration --
in production something like a thread pool should be used.